pyproject.toml
[tool.poetry]
name = "frameserver"
version = "0.1.0"
description = ""
authors = ["Oscar Moll <orm@csail.mit.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "~0.75"
uvicorn = "~0.17"
Pillow = "~9"
av = "~9"
pandas = "~1"
tqdm = "^4.64.0"
pyarrow = "^7.0.0"
scikit-video = "^1.1.11"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
torch = "^1.11.0"
torchvision = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
